ProductUpgraderInfo: undo partial classes
The GVFS/GVFS.Common/ProductUpgraderInfo.cs file was split into
two and GVFS/GVFS.Common/ProductUpgraderInfo.Shared.cs was created
in commit d5d9e7d3c5f822436d4ba7283e18581fc4d528f7, in order to
share the IsLocalUpgradeAvailable() method and some constant path
strings with the GVFS.Hooks project.

Given that there is no longer a corresponding Scalar version of
the GVFS.Hooks project, there is no need to split out certain
(static) methods from the others in any classes.  (The GVFS.Hooks
project was removed in commit
f12effc371d7b1457ab58c6d44959632404e5998 in PR #23.)

Further, the IsLocalUpgradeAvailable() method's only caller was
in the GVFS.Hooks project, so we can just drop that method and
one private one as well.  We do merge another private method
and the path constants into ProductUpgraderInfo.cs and then
remove the ProductUpgraderInfo.Shared.cs file.
