+++ Scalar.Common/ProductUpgraderInfo.cs
@@ -5,8 +5,20 @@ using System.IO;
 
 namespace Scalar.Common
 {
-    public partial class ProductUpgraderInfo
+    public class ProductUpgraderInfo
     {
+        public const string UpgradeDirectoryName = "Scalar.Upgrade";
+        public const string LogDirectory = "UpgraderLogs";
+        public const string DownloadDirectory = "Downloads";
+
+        /// <summary>
+        /// This is the name of the directory that the Upgrader Application is copied to
+        /// when running upgrade. It is called "Tools", as this is what the directory was
+        /// originally named, but has since been renamed in code to be more descriptive.
+        /// </summary>
+        public const string ApplicationDirectory = "Tools";
+        public const string HighestAvailableVersionFileName = "HighestAvailableVersion";
+
         private ITracer tracer;
         private PhysicalFileSystem fileSystem;
 
@@ -97,6 +109,11 @@ namespace Scalar.Common
                     this.tracer.RelatedInfo($"{nameof(this.RecordHighestAvailableVersion)}: Created upgrade reminder marker file");
                 }
             }
+        }
+
+        private static string GetHighestAvailableVersionFilePath(string highestAvailableVersionDirectory)
+        {
+            return Path.Combine(highestAvailableVersionDirectory, HighestAvailableVersionFileName);
         }
     }
 }
